Run our CI on Ubuntu 20.04 as well
Since we're limited to running 5 jobs simultaneously and Ubuntu 16.04 is
nearing its End Of Life, it's unlikely new installations will use it, so
I've removed the job installing Ubuntu 16.04 in a staging environment.
We're still testing its installation on production environments, so it
shouldn't be a big issue.

We're using Python 3.8 in this case because it's the default version in
Ubuntu 20.04.